+++ hydra/symmetries/operations/fermi_sign.h
@@ -4,11 +4,11 @@
 
 namespace hydra::symmetries {
 
-std::vector<int> fermi_work(int n_sites);
+std::vector<int64_t> fermi_work(int64_t n_sites);
 
 template <class bit_t>
 bool fermi_bool_of_permutation(bit_t state, Permutation const &permutation,
-                               std::vector<int> &work);
+                               std::vector<int64_t> &work);
 // "work" needs to be allocated of size n_sites + 4
 
 template <class bit_t>
@@ -20,7 +20,7 @@ inline bool fermi_bool_of_permutation(bit_t state,
 
 template <class bit_t>
 double fermi_sign_of_permutation(bit_t state, Permutation const &permutation,
-                                 std::vector<int> &work);
+                                 std::vector<int64_t> &work);
 // "work" needs to be allocated of size n_sites
 
 template <class bit_t>
@@ -30,17 +30,17 @@ inline bool fermi_sign_of_permutation(bit_t state,
   return fermi_sign_of_permutation(state, permutation, work);
 }
 
-std::vector<int> fermi_work_sort(int n_sites);
+std::vector<int64_t> fermi_work_sort(int64_t n_sites);
 
 template <class bit_t>
 bool fermi_bool_of_permutation_sort(bit_t state, Permutation const &permutation,
-                                    std::vector<int> &work);
+                                    std::vector<int64_t> &work);
 // "work" needs to be allocated of size 2*n_sites
 
 template <class bit_t>
 double fermi_sign_of_permutation_sort(bit_t state,
                                       Permutation const &permutation,
-                                      std::vector<int> &work);
+                                      std::vector<int64_t> &work);
 // "work" needs to be allocated of size 2*n_sites
 
 } // namespace hydra::symmetries
